Wald Test for Two Q-vectors

This function flexibly provides the Wald test for any two q-vectors of a given item in the Q-matrix, but requires that the two q-vectors differ by only one attribute. Additionally, this function needs to accept a CDM.obj.

Wald.test(CDM.obj, q1, q2, i = 1)

Arguments

CDM.obj

An object of class CDM.obj. @seealso CDM.

q1

A q-vector

q2

Another q-vector

i

the item needed to be validated

Value

An object of class htest containing the following components:

statistic

The statistic of the Wald test.

parameter

the degrees of freedom for the Wald-statistic.

p.value

The p value

Details

$$ Wald = \left[\boldsymbol{R} \times \boldsymbol{P}_{i}(\boldsymbol{\alpha})\right]^{'} (\boldsymbol{R} \times \boldsymbol{V}_{i} \times \boldsymbol{R})^{-1} \left[\boldsymbol{R} \times P_{i}(\boldsymbol{\alpha})\right] $$

Examples

library(Qval)
set.seed(123)

K <- 3
I <- 20
N <- 500
IQ <- list(
  P0 = runif(I, 0.0, 0.2),
  P1 = runif(I, 0.8, 1.0)
)
Q <- sim.Q(K, I)
data <- sim.data(Q = Q, N = N, IQ = IQ, model = "GDINA", distribute = "horder")
#> distribute =  horder 
#> model =  GDINA 
#>  number of attributes:  3 
#>  number of items:  20 
#>  num of examinees:  500 
#>  average of P0 =  0.11 
#>  average of P1 =  0.915 
#> theta_mean =  0.03 , theta_sd = 0.969 
#>  a =  1.5 1.5 1.5 
#>  b =  -1.5 1.5 0 

CDM.obj <- CDM(data$dat, Q)
#> Iter = 1  Max. abs. change = 0.40236  Deviance  = 10482.49                                                                                  Iter = 2  Max. abs. change = 0.22441  Deviance  = 8833.97                                                                                  Iter = 3  Max. abs. change = 0.06646  Deviance  = 8785.59                                                                                  Iter = 4  Max. abs. change = 0.01513  Deviance  = 8776.51                                                                                  Iter = 5  Max. abs. change = 0.00506  Deviance  = 8774.73                                                                                  Iter = 6  Max. abs. change = 0.00254  Deviance  = 8774.30                                                                                  Iter = 7  Max. abs. change = 0.00136  Deviance  = 8774.18                                                                                  Iter = 8  Max. abs. change = 0.00073  Deviance  = 8774.14                                                                                  Iter = 9  Max. abs. change = 0.00039  Deviance  = 8774.12                                                                                  Iter = 10  Max. abs. change = 0.00021  Deviance  = 8774.12                                                                                  Iter = 11  Max. abs. change = 0.00011  Deviance  = 8774.12                                                                                  Iter = 12  Max. abs. change = 0.00005  Deviance  = 8774.12                                                                                  

q1 <- c(1, 0, 0)
q2 <- c(1, 1, 0)

## Discuss whether there is a significant difference when 
## the q-vector of the 2nd item in the Q-matrix is q1 or q2.
Wald.test.obj <- Wald.test(CDM.obj, q1, q2, i=2)

print(Wald.test.obj)
#> 
#> 	Wald test for two q-vectors
#> 
#> data:  [100] vs. [110] for item 2
#> Wald = 6.4153, df = 2, p-value = 0.04045
#>