Iterative Attribute Hierarchy Exploration Methods for Cognitive Diagnosis Models

This function implements an exploratory method for attribute hierarchy structure (Zhang et al., 2025). The procedure is iterative: in each step any structural parameter that is

less than or equal to the pre-specified threshold eps, or

not greater than 0 (tested by z-statistic under Bonferroni correction and standard erros are from XPD information matrix)

is fixed to zero and the remaining parameters are re-estimated. Upon convergence, all structural parameters are both greater than eps and significantly larger than 0. The attribute hierarchy is then inferred from the set of parameters that remain significantly positive.

att.hierarchy(
  Y,
  Q,
  model = "GDINA",
  mono.constraint = FALSE,
  maxitr = 20,
  eps = 1e-07,
  alpha.level = 0.01,
  verbose = TRUE
)

Arguments

Y

A required \(N\) × \(I\) matrix or data.frame consisting of the responses of N individuals to \(I\) items. Missing values need to be coded as NA.

Q

A required binary \(I\) × \(K\) matrix containing the attributes not required or required master the items. The ith row of the matrix is a binary indicator vector indicating which attributes are not required (coded by 0) and which attributes are required (coded by 1) to master item \(i\).

model

Type of model to be fitted; can be "GDINA", "LCDM", "DINA", "DINO", "ACDM", "LLM", or "rRUM". Default = "GDINA".

mono.constraint

Logical indicating whether monotonicity constraints should be fulfilled in estimation. Default = FALSE.

maxitr

Number of max iterations. Default = 20.

eps

Cut-off points of the minimum vallue of structural parameters.

alpha.level

alpha level for the z-statistic test under Bonferroni correction. Default = 0.01.

verbose

Logical indicating to print iterative information or not. Default is TRUE

Value

An object of class att.hierarchy containing the following components:

statistic

A 4-column matrix for each structural parameter that is significantly larger than 0: the parameter estimate, its standard error (SE), the corresponding z-statistic, and the p-values after Bonferroni correction.

noSig

A logical scalar: TRUE if, during iteration, all structural parameters are not greater than 0; otherwise FALSE.

isNonverge

A logical scalar: TRUE if convergence was achieved within maxitr iterations; FALSE if the algorithm did not converged.

pattern

The attribute mastery pattern matrix that contains every possible attribute mastery pattern.

arguments

A list that stores all input arguments supplied by the user.

References

Zhang, X., Jiang, Y., Xin, T., & Liu, Y. (2025). Iterative Attribute Hierarchy Exploration Methods for Cognitive Diagnosis Models. Journal of Educational and Behavioral Statistics, 50(4), 682-713. https://doi.org/10.3102/10769986241268906

Author

Haijiang Qin <Haijiang133@outlook.com>