Calculate \(\beta\)

The function is able to calculate the \(\beta\) index for all items after fitting CDM or directly.

get.beta(
  Y = NULL,
  Q = NULL,
  att.str = NULL,
  CDM.obj = NULL,
  mono.constraint = FALSE,
  model = "GDINA"
)

Arguments

Y

A required \(N\) × \(I\) matrix or data.frame consisting of the responses of N individuals to \(N\) × \(I\) items. Missing values need to be coded as NA.

Q

A required binary \(I\) × \(K\) matrix containing the attributes not required or required master the items. The ith row of the matrix is a binary indicator vector indicating which attributes are not required (coded by 0) and which attributes are required (coded by 1) to master item \(i\).

att.str

Specify attribute structures. NULL, by default, means there is no structure. Attribute structure needs be specified as a list - which will be internally handled by att.structure function. See examples. It can also be a matrix giving all permissible attribute profiles.

CDM.obj

An object of class CDM.obj. Can be NULL, but when it is not NULL, it enables rapid validation of the Q-matrix without the need for parameter estimation. @seealso CDM.

mono.constraint

Logical indicating whether monotonicity constraints should be fulfilled in estimation. Default = FALSE.

model

Type of model to be fitted; can be "GDINA", "LCDM", "DINA", "DINO", "ACDM", "LLM", or "rRUM". Default = "GDINA".

Value

An object of class matrix, which consisted of \(\beta\) index for each item and each possible attribute mastery pattern.

Details

For item \(i\) with the q-vector of the \(c\)-th (\(c = 1, 2, ..., 2^{K}\)) type, the \(\beta\) index is computed as follows:

$$ \beta_{ic} = \sum_{l=1}^{2^K} \left| \frac{r_{li}}{n_l} P_{ic}(\boldsymbol{\alpha}_{l}) - \left(1 - \frac{r_{li}}{n_l}\right) \left[1 - P_{ic}(\boldsymbol{\alpha}_{l})\right] \right| = \sum_{l=1}^{2^K} \left| \frac{r_{li}}{n_l} - \left[1 - P_{ic}(\boldsymbol{\alpha}_{l}) \right] \right| $$

In the formula, \(r_{li}\) represents the number of examinees in attribute mastery pattern \(\boldsymbol{\alpha}_{l}\) who correctly answered item \(i\), while \(n_l\) is the total number of examinees in attribute mastery pattern \(\boldsymbol{\alpha}_{l}\). \(P_{ic}(\boldsymbol{\alpha}_{l})\) denotes the probability that an examinee in attribute mastery pattern \(\boldsymbol{\alpha}_{l}\) answers item \(i\) correctly when the q-vector for item \(i\) is of the \(c\)-th type. In fact, \(\frac{r_{li}}{n_l}\) is the observed probability that an examinee in attribute mastery pattern \(\boldsymbol{\alpha}_{l}\) answers item \(i\) correctly, and \(\beta_{jc}\) represents the difference between the actual proportion of correct answers for item \(i\) in each attribute mastery pattern and the expected probability of answering the item incorrectly in that state. Therefore, to some extent, \(\beta_{jc}\) can be considered as a measure of discriminability.

References

Li, J., & Chen, P. (2024). A new Q-matrix validation method based on signal detection theory. British Journal of Mathematical and Statistical Psychology, 00, 1–33. DOI: 10.1111/bmsp.12371

See also

Author

Haijiang Qin <Haijiang133@outlook.com>

Examples

# \donttest{
library(Qval)

set.seed(123)

## generate Q-matrix and data
K <- 4
I <- 20
Q <- sim.Q(K, I)
IQ <- list(
  P0 = runif(I, 0.0, 0.3),
  P1 = runif(I, 0.7, 0.9)
)

model <- "GDINA"
data <- sim.data(Q = Q, N = 500, IQ = IQ, model = model, distribute = "horder")
#> distribute =  horder 
#> model =  GDINA 
#>  number of attributes:  4 
#>  number of items:  20 
#>  num of examinees:  500 
#>  average of P0 =  0.128 
#>  average of P1 =  0.798 
#> theta_mean =  -0.06 , theta_sd = 0.996 
#>  a =  1.5 1.5 1.5 1.5 
#>  b =  0.5 1.5 -1.5 -0.5 

## calculate beta directly
beta <-get.beta(Y = data$dat, Q = Q, model = model)
#> Iter = 1  Max. abs. change = 0.38660  Deviance  = 12248.06                                                                                  Iter = 2  Max. abs. change = 0.10531  Deviance  = 11181.61                                                                                  Iter = 3  Max. abs. change = 0.06712  Deviance  = 11126.61                                                                                  Iter = 4  Max. abs. change = 0.03723  Deviance  = 11109.01                                                                                  Iter = 5  Max. abs. change = 0.02670  Deviance  = 11101.45                                                                                  Iter = 6  Max. abs. change = 0.02183  Deviance  = 11097.63                                                                                  Iter = 7  Max. abs. change = 0.01829  Deviance  = 11095.36                                                                                  Iter = 8  Max. abs. change = 0.01557  Deviance  = 11093.82                                                                                  Iter = 9  Max. abs. change = 0.01336  Deviance  = 11092.67                                                                                  Iter = 10  Max. abs. change = 0.01150  Deviance  = 11091.75                                                                                  Iter = 11  Max. abs. change = 0.01016  Deviance  = 11090.99                                                                                  Iter = 12  Max. abs. change = 0.01009  Deviance  = 11090.35                                                                                  Iter = 13  Max. abs. change = 0.00997  Deviance  = 11089.80                                                                                  Iter = 14  Max. abs. change = 0.00976  Deviance  = 11089.32                                                                                  Iter = 15  Max. abs. change = 0.00944  Deviance  = 11088.90                                                                                  Iter = 16  Max. abs. change = 0.00899  Deviance  = 11088.54                                                                                  Iter = 17  Max. abs. change = 0.00844  Deviance  = 11088.23                                                                                  Iter = 18  Max. abs. change = 0.00777  Deviance  = 11087.95                                                                                  Iter = 19  Max. abs. change = 0.00703  Deviance  = 11087.72                                                                                  Iter = 20  Max. abs. change = 0.00624  Deviance  = 11087.52                                                                                  Iter = 21  Max. abs. change = 0.00543  Deviance  = 11087.35                                                                                  Iter = 22  Max. abs. change = 0.00480  Deviance  = 11087.21                                                                                  Iter = 23  Max. abs. change = 0.00428  Deviance  = 11087.09                                                                                  Iter = 24  Max. abs. change = 0.00379  Deviance  = 11086.99                                                                                  Iter = 25  Max. abs. change = 0.00332  Deviance  = 11086.91                                                                                  Iter = 26  Max. abs. change = 0.00288  Deviance  = 11086.84                                                                                  Iter = 27  Max. abs. change = 0.00275  Deviance  = 11086.79                                                                                  Iter = 28  Max. abs. change = 0.00272  Deviance  = 11086.75                                                                                  Iter = 29  Max. abs. change = 0.00267  Deviance  = 11086.72                                                                                  Iter = 30  Max. abs. change = 0.00262  Deviance  = 11086.69                                                                                  Iter = 31  Max. abs. change = 0.00257  Deviance  = 11086.67                                                                                  Iter = 32  Max. abs. change = 0.00251  Deviance  = 11086.65                                                                                  Iter = 33  Max. abs. change = 0.00245  Deviance  = 11086.64                                                                                  Iter = 34  Max. abs. change = 0.00238  Deviance  = 11086.63                                                                                  Iter = 35  Max. abs. change = 0.00232  Deviance  = 11086.62                                                                                  Iter = 36  Max. abs. change = 0.00225  Deviance  = 11086.61                                                                                  Iter = 37  Max. abs. change = 0.00219  Deviance  = 11086.60                                                                                  Iter = 38  Max. abs. change = 0.00213  Deviance  = 11086.60                                                                                  Iter = 39  Max. abs. change = 0.00206  Deviance  = 11086.59                                                                                  Iter = 40  Max. abs. change = 0.00200  Deviance  = 11086.59                                                                                  Iter = 41  Max. abs. change = 0.00194  Deviance  = 11086.59                                                                                  Iter = 42  Max. abs. change = 0.00188  Deviance  = 11086.58                                                                                  Iter = 43  Max. abs. change = 0.00183  Deviance  = 11086.58                                                                                  Iter = 44  Max. abs. change = 0.00177  Deviance  = 11086.58                                                                                  Iter = 45  Max. abs. change = 0.00172  Deviance  = 11086.58                                                                                  Iter = 46  Max. abs. change = 0.00167  Deviance  = 11086.57                                                                                  Iter = 47  Max. abs. change = 0.00162  Deviance  = 11086.57                                                                                  Iter = 48  Max. abs. change = 0.00157  Deviance  = 11086.57                                                                                  Iter = 49  Max. abs. change = 0.00153  Deviance  = 11086.57                                                                                  Iter = 50  Max. abs. change = 0.00148  Deviance  = 11086.57                                                                                  Iter = 51  Max. abs. change = 0.00144  Deviance  = 11086.57                                                                                  Iter = 52  Max. abs. change = 0.00140  Deviance  = 11086.57                                                                                  Iter = 53  Max. abs. change = 0.00136  Deviance  = 11086.56                                                                                  Iter = 54  Max. abs. change = 0.00132  Deviance  = 11086.56                                                                                  Iter = 55  Max. abs. change = 0.00128  Deviance  = 11086.56                                                                                  Iter = 56  Max. abs. change = 0.00124  Deviance  = 11086.56                                                                                  Iter = 57  Max. abs. change = 0.00121  Deviance  = 11086.56                                                                                  Iter = 58  Max. abs. change = 0.00117  Deviance  = 11086.56                                                                                  Iter = 59  Max. abs. change = 0.00114  Deviance  = 11086.56                                                                                  Iter = 60  Max. abs. change = 0.00110  Deviance  = 11086.56                                                                                  Iter = 61  Max. abs. change = 0.00107  Deviance  = 11086.56                                                                                  Iter = 62  Max. abs. change = 0.00104  Deviance  = 11086.55                                                                                  Iter = 63  Max. abs. change = 0.00101  Deviance  = 11086.55                                                                                  Iter = 64  Max. abs. change = 0.00098  Deviance  = 11086.55                                                                                  Iter = 65  Max. abs. change = 0.00095  Deviance  = 11086.55                                                                                  Iter = 66  Max. abs. change = 0.00092  Deviance  = 11086.55                                                                                  Iter = 67  Max. abs. change = 0.00089  Deviance  = 11086.55                                                                                  Iter = 68  Max. abs. change = 0.00086  Deviance  = 11086.55                                                                                  Iter = 69  Max. abs. change = 0.00083  Deviance  = 11086.55                                                                                  Iter = 70  Max. abs. change = 0.00081  Deviance  = 11086.55                                                                                  Iter = 71  Max. abs. change = 0.00078  Deviance  = 11086.55                                                                                  Iter = 72  Max. abs. change = 0.00075  Deviance  = 11086.55                                                                                  Iter = 73  Max. abs. change = 0.00073  Deviance  = 11086.55                                                                                  Iter = 74  Max. abs. change = 0.00070  Deviance  = 11086.55                                                                                  Iter = 75  Max. abs. change = 0.00068  Deviance  = 11086.55                                                                                  Iter = 76  Max. abs. change = 0.00065  Deviance  = 11086.55                                                                                  Iter = 77  Max. abs. change = 0.00063  Deviance  = 11086.55                                                                                  Iter = 78  Max. abs. change = 0.00061  Deviance  = 11086.54                                                                                  Iter = 79  Max. abs. change = 0.00058  Deviance  = 11086.54                                                                                  Iter = 80  Max. abs. change = 0.00056  Deviance  = 11086.54                                                                                  Iter = 81  Max. abs. change = 0.00054  Deviance  = 11086.54                                                                                  Iter = 82  Max. abs. change = 0.00052  Deviance  = 11086.54                                                                                  Iter = 83  Max. abs. change = 0.00050  Deviance  = 11086.54                                                                                  Iter = 84  Max. abs. change = 0.00048  Deviance  = 11086.54                                                                                  Iter = 85  Max. abs. change = 0.00046  Deviance  = 11086.54                                                                                  Iter = 86  Max. abs. change = 0.00044  Deviance  = 11086.54                                                                                  Iter = 87  Max. abs. change = 0.00042  Deviance  = 11086.54                                                                                  Iter = 88  Max. abs. change = 0.00041  Deviance  = 11086.54                                                                                  Iter = 89  Max. abs. change = 0.00039  Deviance  = 11086.54                                                                                  Iter = 90  Max. abs. change = 0.00037  Deviance  = 11086.54                                                                                  Iter = 91  Max. abs. change = 0.00036  Deviance  = 11086.54                                                                                  Iter = 92  Max. abs. change = 0.00034  Deviance  = 11086.54                                                                                  Iter = 93  Max. abs. change = 0.00033  Deviance  = 11086.54                                                                                  Iter = 94  Max. abs. change = 0.00031  Deviance  = 11086.54                                                                                  Iter = 95  Max. abs. change = 0.00030  Deviance  = 11086.54                                                                                  Iter = 96  Max. abs. change = 0.00028  Deviance  = 11086.54                                                                                  Iter = 97  Max. abs. change = 0.00027  Deviance  = 11086.54                                                                                  Iter = 98  Max. abs. change = 0.00026  Deviance  = 11086.54                                                                                  Iter = 99  Max. abs. change = 0.00025  Deviance  = 11086.54                                                                                  Iter = 100  Max. abs. change = 0.00024  Deviance  = 11086.54                                                                                  Iter = 101  Max. abs. change = 0.00022  Deviance  = 11086.54                                                                                  Iter = 102  Max. abs. change = 0.00021  Deviance  = 11086.54                                                                                  Iter = 103  Max. abs. change = 0.00020  Deviance  = 11086.54                                                                                  Iter = 104  Max. abs. change = 0.00019  Deviance  = 11086.54                                                                                  Iter = 105  Max. abs. change = 0.00018  Deviance  = 11086.54                                                                                  Iter = 106  Max. abs. change = 0.00018  Deviance  = 11086.54                                                                                  Iter = 107  Max. abs. change = 0.00017  Deviance  = 11086.54                                                                                  Iter = 108  Max. abs. change = 0.00016  Deviance  = 11086.54                                                                                  Iter = 109  Max. abs. change = 0.00015  Deviance  = 11086.54                                                                                  Iter = 110  Max. abs. change = 0.00014  Deviance  = 11086.54                                                                                  Iter = 111  Max. abs. change = 0.00014  Deviance  = 11086.54                                                                                  Iter = 112  Max. abs. change = 0.00013  Deviance  = 11086.54                                                                                  Iter = 113  Max. abs. change = 0.00012  Deviance  = 11086.54                                                                                  Iter = 114  Max. abs. change = 0.00012  Deviance  = 11086.54                                                                                  Iter = 115  Max. abs. change = 0.00011  Deviance  = 11086.54                                                                                  Iter = 116  Max. abs. change = 0.00011  Deviance  = 11086.54                                                                                  Iter = 117  Max. abs. change = 0.00010  Deviance  = 11086.54                                                                                  Iter = 118  Max. abs. change = 0.00010  Deviance  = 11086.54                                                                                  
print(beta)
#>         0000      1000     0100      0010      0001      1100      1010
#> item 1     0 12.527727 6.358972  6.670214  6.552147 12.649119 12.548779
#> item 2     0  4.707686 4.805865  7.887206  4.827853  4.877648  7.790552
#> item 3     0  6.007082 5.962630 11.083919  5.817212  6.023786 11.054038
#> item 4     0  6.616149 6.429311 12.031596  6.331844  6.612520 11.961236
#> item 5     0  4.409362 4.455982  4.430382  8.285693  4.479183  4.411032
#> item 6     0  7.272588 7.201482 13.368557  7.049923  7.339736 13.470615
#> item 7     0  6.071023 5.986996 10.713982  6.043961  6.074383 10.747575
#> item 8     0  6.356890 4.405264  6.587027  4.600520  6.170474  8.374300
#> item 9     0  5.970399 4.737318  4.840204  3.881811  6.228024  5.874771
#> item 10    0  5.965935 5.106570  6.591726  5.945526  5.936155  7.836351
#> item 11    0  6.455260 6.542014  6.289457 11.504399  6.579281  6.348872
#> item 12    0 11.565140 6.553855  6.234117  6.559488 11.550940 11.557522
#> item 13    0  6.202124 3.555491  3.858179  3.722392  6.182562  5.857571
#> item 14    0  5.141324 3.080738  3.658573  2.836921  5.107840  4.794495
#> item 15    0  4.597051 4.280770  3.213487  4.655373  5.447809  4.344889
#> item 16    0  6.478585 3.868452  3.791174  4.155395  6.426057  6.508694
#> item 17    0  5.314695 7.443550  5.009076  5.533609  7.257750  5.355726
#> item 18    0  5.174471 9.204049  5.042170  5.315518  9.166368  5.075680
#> item 19    0  6.400072 6.279604  9.616281  6.660542  6.489507  9.543705
#> item 20    0  4.739066 8.683251  4.965112  4.879851  8.715230  4.773643
#>              1001      0110      0101      0011      1110      1101      1011
#> item 1  12.529287  6.569089  6.481237  6.687773 12.673875 12.662218 12.574348
#> item 2   4.954292  8.818432  4.990899  7.471134  8.686694  5.303359  7.492397
#> item 3   6.007267 11.148447  5.883571 11.115449 11.243742  6.024432 11.071739
#> item 4   6.616414 12.199613  6.246872 12.027932 12.179587  6.769262 11.966216
#> item 5   8.276556  4.444049  8.161379  8.269638  4.390555  8.176133  8.131217
#> item 6   7.272939 13.424433  7.094848 13.394183 13.535782  7.390438 13.484083
#> item 7   6.170483 10.750257  6.039117 10.614278 10.828102  6.151650 10.687193
#> item 8   6.368025  6.464822  4.432688  6.592052  8.209456  6.220541  8.427957
#> item 9   5.858782  5.713194  4.654576  4.792043  7.386009  6.344011  5.740506
#> item 10  6.359140  6.519868  6.226253  7.093199  7.933930  6.713593  8.054612
#> item 11 11.512609  6.376044 11.502417 11.634465  6.749914 11.461080 11.671445
#> item 12 11.538174  6.354960  6.741973  6.418011 11.530548 11.622188 11.603220
#> item 13  6.330339  3.839737  3.822262  3.990550  5.748591  6.402797  6.004828
#> item 14  5.132207  3.633435  3.086874  3.658142  4.818689  5.225482  4.818454
#> item 15  5.380013  4.426031  4.766321  4.387786  5.127774  5.353639  5.074417
#> item 16  6.573114  3.850395  4.208478  4.167015  6.503793  6.595944  6.663296
#> item 17  6.204606  7.424443  7.650946  5.579625  7.542292  8.544501  6.438509
#> item 18  5.488626  9.223433  9.295840  5.186586  9.238719  9.324354  5.997775
#> item 19  7.041727  9.575667  6.637693  9.504269 10.626007  7.675357  9.623505
#> item 20  4.883998  8.532811  8.614262  4.909266  8.450907  8.663136  4.883410
#>              0111      1111
#> item 1   6.859112 12.771521
#> item 2   8.749377  8.630208
#> item 3  11.071170 11.310109
#> item 4  12.163319 12.172053
#> item 5   8.016842  8.136382
#> item 6  13.453244 13.612043
#> item 7  10.663114 10.715242
#> item 8   6.491115  8.285429
#> item 9   6.074754  7.418073
#> item 10  7.313084  8.708345
#> item 11 11.670328 11.660831
#> item 12  7.407881 11.786779
#> item 13  4.676216  6.058610
#> item 14  3.718746  5.131611
#> item 15  4.823958  5.308987
#> item 16  4.081800  6.855066
#> item 17  8.398387  8.398041
#> item 18  9.264553  9.629526
#> item 19  9.630452 10.106430
#> item 20  8.343755  8.322236

## calculate beta after fitting CDM
CDM.obj <- CDM(data$dat, Q, model=model)
#> Iter = 1  Max. abs. change = 0.38660  Deviance  = 12248.06                                                                                  Iter = 2  Max. abs. change = 0.10531  Deviance  = 11181.61                                                                                  Iter = 3  Max. abs. change = 0.06712  Deviance  = 11126.61                                                                                  Iter = 4  Max. abs. change = 0.03723  Deviance  = 11109.01                                                                                  Iter = 5  Max. abs. change = 0.02670  Deviance  = 11101.45                                                                                  Iter = 6  Max. abs. change = 0.02183  Deviance  = 11097.63                                                                                  Iter = 7  Max. abs. change = 0.01829  Deviance  = 11095.36                                                                                  Iter = 8  Max. abs. change = 0.01557  Deviance  = 11093.82                                                                                  Iter = 9  Max. abs. change = 0.01336  Deviance  = 11092.67                                                                                  Iter = 10  Max. abs. change = 0.01150  Deviance  = 11091.75                                                                                  Iter = 11  Max. abs. change = 0.01016  Deviance  = 11090.99                                                                                  Iter = 12  Max. abs. change = 0.01009  Deviance  = 11090.35                                                                                  Iter = 13  Max. abs. change = 0.00997  Deviance  = 11089.80                                                                                  Iter = 14  Max. abs. change = 0.00976  Deviance  = 11089.32                                                                                  Iter = 15  Max. abs. change = 0.00944  Deviance  = 11088.90                                                                                  Iter = 16  Max. abs. change = 0.00899  Deviance  = 11088.54                                                                                  Iter = 17  Max. abs. change = 0.00844  Deviance  = 11088.23                                                                                  Iter = 18  Max. abs. change = 0.00777  Deviance  = 11087.95                                                                                  Iter = 19  Max. abs. change = 0.00703  Deviance  = 11087.72                                                                                  Iter = 20  Max. abs. change = 0.00624  Deviance  = 11087.52                                                                                  Iter = 21  Max. abs. change = 0.00543  Deviance  = 11087.35                                                                                  Iter = 22  Max. abs. change = 0.00480  Deviance  = 11087.21                                                                                  Iter = 23  Max. abs. change = 0.00428  Deviance  = 11087.09                                                                                  Iter = 24  Max. abs. change = 0.00379  Deviance  = 11086.99                                                                                  Iter = 25  Max. abs. change = 0.00332  Deviance  = 11086.91                                                                                  Iter = 26  Max. abs. change = 0.00288  Deviance  = 11086.84                                                                                  Iter = 27  Max. abs. change = 0.00275  Deviance  = 11086.79                                                                                  Iter = 28  Max. abs. change = 0.00272  Deviance  = 11086.75                                                                                  Iter = 29  Max. abs. change = 0.00267  Deviance  = 11086.72                                                                                  Iter = 30  Max. abs. change = 0.00262  Deviance  = 11086.69                                                                                  Iter = 31  Max. abs. change = 0.00257  Deviance  = 11086.67                                                                                  Iter = 32  Max. abs. change = 0.00251  Deviance  = 11086.65                                                                                  Iter = 33  Max. abs. change = 0.00245  Deviance  = 11086.64                                                                                  Iter = 34  Max. abs. change = 0.00238  Deviance  = 11086.63                                                                                  Iter = 35  Max. abs. change = 0.00232  Deviance  = 11086.62                                                                                  Iter = 36  Max. abs. change = 0.00225  Deviance  = 11086.61                                                                                  Iter = 37  Max. abs. change = 0.00219  Deviance  = 11086.60                                                                                  Iter = 38  Max. abs. change = 0.00213  Deviance  = 11086.60                                                                                  Iter = 39  Max. abs. change = 0.00206  Deviance  = 11086.59                                                                                  Iter = 40  Max. abs. change = 0.00200  Deviance  = 11086.59                                                                                  Iter = 41  Max. abs. change = 0.00194  Deviance  = 11086.59                                                                                  Iter = 42  Max. abs. change = 0.00188  Deviance  = 11086.58                                                                                  Iter = 43  Max. abs. change = 0.00183  Deviance  = 11086.58                                                                                  Iter = 44  Max. abs. change = 0.00177  Deviance  = 11086.58                                                                                  Iter = 45  Max. abs. change = 0.00172  Deviance  = 11086.58                                                                                  Iter = 46  Max. abs. change = 0.00167  Deviance  = 11086.57                                                                                  Iter = 47  Max. abs. change = 0.00162  Deviance  = 11086.57                                                                                  Iter = 48  Max. abs. change = 0.00157  Deviance  = 11086.57                                                                                  Iter = 49  Max. abs. change = 0.00153  Deviance  = 11086.57                                                                                  Iter = 50  Max. abs. change = 0.00148  Deviance  = 11086.57                                                                                  Iter = 51  Max. abs. change = 0.00144  Deviance  = 11086.57                                                                                  Iter = 52  Max. abs. change = 0.00140  Deviance  = 11086.57                                                                                  Iter = 53  Max. abs. change = 0.00136  Deviance  = 11086.56                                                                                  Iter = 54  Max. abs. change = 0.00132  Deviance  = 11086.56                                                                                  Iter = 55  Max. abs. change = 0.00128  Deviance  = 11086.56                                                                                  Iter = 56  Max. abs. change = 0.00124  Deviance  = 11086.56                                                                                  Iter = 57  Max. abs. change = 0.00121  Deviance  = 11086.56                                                                                  Iter = 58  Max. abs. change = 0.00117  Deviance  = 11086.56                                                                                  Iter = 59  Max. abs. change = 0.00114  Deviance  = 11086.56                                                                                  Iter = 60  Max. abs. change = 0.00110  Deviance  = 11086.56                                                                                  Iter = 61  Max. abs. change = 0.00107  Deviance  = 11086.56                                                                                  Iter = 62  Max. abs. change = 0.00104  Deviance  = 11086.55                                                                                  Iter = 63  Max. abs. change = 0.00101  Deviance  = 11086.55                                                                                  Iter = 64  Max. abs. change = 0.00098  Deviance  = 11086.55                                                                                  Iter = 65  Max. abs. change = 0.00095  Deviance  = 11086.55                                                                                  Iter = 66  Max. abs. change = 0.00092  Deviance  = 11086.55                                                                                  Iter = 67  Max. abs. change = 0.00089  Deviance  = 11086.55                                                                                  Iter = 68  Max. abs. change = 0.00086  Deviance  = 11086.55                                                                                  Iter = 69  Max. abs. change = 0.00083  Deviance  = 11086.55                                                                                  Iter = 70  Max. abs. change = 0.00081  Deviance  = 11086.55                                                                                  Iter = 71  Max. abs. change = 0.00078  Deviance  = 11086.55                                                                                  Iter = 72  Max. abs. change = 0.00075  Deviance  = 11086.55                                                                                  Iter = 73  Max. abs. change = 0.00073  Deviance  = 11086.55                                                                                  Iter = 74  Max. abs. change = 0.00070  Deviance  = 11086.55                                                                                  Iter = 75  Max. abs. change = 0.00068  Deviance  = 11086.55                                                                                  Iter = 76  Max. abs. change = 0.00065  Deviance  = 11086.55                                                                                  Iter = 77  Max. abs. change = 0.00063  Deviance  = 11086.55                                                                                  Iter = 78  Max. abs. change = 0.00061  Deviance  = 11086.54                                                                                  Iter = 79  Max. abs. change = 0.00058  Deviance  = 11086.54                                                                                  Iter = 80  Max. abs. change = 0.00056  Deviance  = 11086.54                                                                                  Iter = 81  Max. abs. change = 0.00054  Deviance  = 11086.54                                                                                  Iter = 82  Max. abs. change = 0.00052  Deviance  = 11086.54                                                                                  Iter = 83  Max. abs. change = 0.00050  Deviance  = 11086.54                                                                                  Iter = 84  Max. abs. change = 0.00048  Deviance  = 11086.54                                                                                  Iter = 85  Max. abs. change = 0.00046  Deviance  = 11086.54                                                                                  Iter = 86  Max. abs. change = 0.00044  Deviance  = 11086.54                                                                                  Iter = 87  Max. abs. change = 0.00042  Deviance  = 11086.54                                                                                  Iter = 88  Max. abs. change = 0.00041  Deviance  = 11086.54                                                                                  Iter = 89  Max. abs. change = 0.00039  Deviance  = 11086.54                                                                                  Iter = 90  Max. abs. change = 0.00037  Deviance  = 11086.54                                                                                  Iter = 91  Max. abs. change = 0.00036  Deviance  = 11086.54                                                                                  Iter = 92  Max. abs. change = 0.00034  Deviance  = 11086.54                                                                                  Iter = 93  Max. abs. change = 0.00033  Deviance  = 11086.54                                                                                  Iter = 94  Max. abs. change = 0.00031  Deviance  = 11086.54                                                                                  Iter = 95  Max. abs. change = 0.00030  Deviance  = 11086.54                                                                                  Iter = 96  Max. abs. change = 0.00028  Deviance  = 11086.54                                                                                  Iter = 97  Max. abs. change = 0.00027  Deviance  = 11086.54                                                                                  Iter = 98  Max. abs. change = 0.00026  Deviance  = 11086.54                                                                                  Iter = 99  Max. abs. change = 0.00025  Deviance  = 11086.54                                                                                  Iter = 100  Max. abs. change = 0.00024  Deviance  = 11086.54                                                                                  Iter = 101  Max. abs. change = 0.00022  Deviance  = 11086.54                                                                                  Iter = 102  Max. abs. change = 0.00021  Deviance  = 11086.54                                                                                  Iter = 103  Max. abs. change = 0.00020  Deviance  = 11086.54                                                                                  Iter = 104  Max. abs. change = 0.00019  Deviance  = 11086.54                                                                                  Iter = 105  Max. abs. change = 0.00018  Deviance  = 11086.54                                                                                  Iter = 106  Max. abs. change = 0.00018  Deviance  = 11086.54                                                                                  Iter = 107  Max. abs. change = 0.00017  Deviance  = 11086.54                                                                                  Iter = 108  Max. abs. change = 0.00016  Deviance  = 11086.54                                                                                  Iter = 109  Max. abs. change = 0.00015  Deviance  = 11086.54                                                                                  Iter = 110  Max. abs. change = 0.00014  Deviance  = 11086.54                                                                                  Iter = 111  Max. abs. change = 0.00014  Deviance  = 11086.54                                                                                  Iter = 112  Max. abs. change = 0.00013  Deviance  = 11086.54                                                                                  Iter = 113  Max. abs. change = 0.00012  Deviance  = 11086.54                                                                                  Iter = 114  Max. abs. change = 0.00012  Deviance  = 11086.54                                                                                  Iter = 115  Max. abs. change = 0.00011  Deviance  = 11086.54                                                                                  Iter = 116  Max. abs. change = 0.00011  Deviance  = 11086.54                                                                                  Iter = 117  Max. abs. change = 0.00010  Deviance  = 11086.54                                                                                  Iter = 118  Max. abs. change = 0.00010  Deviance  = 11086.54                                                                                  
beta <-get.beta(CDM.obj = CDM.obj)
print(beta)
#>         0000      1000     0100      0010      0001      1100      1010
#> item 1     0 12.527727 6.358972  6.670214  6.552147 12.649119 12.548779
#> item 2     0  4.707686 4.805865  7.887206  4.827853  4.877648  7.790552
#> item 3     0  6.007082 5.962630 11.083919  5.817212  6.023786 11.054038
#> item 4     0  6.616149 6.429311 12.031596  6.331844  6.612520 11.961236
#> item 5     0  4.409362 4.455982  4.430382  8.285693  4.479183  4.411032
#> item 6     0  7.272588 7.201482 13.368557  7.049923  7.339736 13.470615
#> item 7     0  6.071023 5.986996 10.713982  6.043961  6.074383 10.747575
#> item 8     0  6.356890 4.405264  6.587027  4.600520  6.170474  8.374300
#> item 9     0  5.970399 4.737318  4.840204  3.881811  6.228024  5.874771
#> item 10    0  5.965935 5.106570  6.591726  5.945526  5.936155  7.836351
#> item 11    0  6.455260 6.542014  6.289457 11.504399  6.579281  6.348872
#> item 12    0 11.565140 6.553855  6.234117  6.559488 11.550940 11.557522
#> item 13    0  6.202124 3.555491  3.858179  3.722392  6.182562  5.857571
#> item 14    0  5.141324 3.080738  3.658573  2.836921  5.107840  4.794495
#> item 15    0  4.597051 4.280770  3.213487  4.655373  5.447809  4.344889
#> item 16    0  6.478585 3.868452  3.791174  4.155395  6.426057  6.508694
#> item 17    0  5.314695 7.443550  5.009076  5.533609  7.257750  5.355726
#> item 18    0  5.174471 9.204049  5.042170  5.315518  9.166368  5.075680
#> item 19    0  6.400072 6.279604  9.616281  6.660542  6.489507  9.543705
#> item 20    0  4.739066 8.683251  4.965112  4.879851  8.715230  4.773643
#>              1001      0110      0101      0011      1110      1101      1011
#> item 1  12.529287  6.569089  6.481237  6.687773 12.673875 12.662218 12.574348
#> item 2   4.954292  8.818432  4.990899  7.471134  8.686694  5.303359  7.492397
#> item 3   6.007267 11.148447  5.883571 11.115449 11.243742  6.024432 11.071739
#> item 4   6.616414 12.199613  6.246872 12.027932 12.179587  6.769262 11.966216
#> item 5   8.276556  4.444049  8.161379  8.269638  4.390555  8.176133  8.131217
#> item 6   7.272939 13.424433  7.094848 13.394183 13.535782  7.390438 13.484083
#> item 7   6.170483 10.750257  6.039117 10.614278 10.828102  6.151650 10.687193
#> item 8   6.368025  6.464822  4.432688  6.592052  8.209456  6.220541  8.427957
#> item 9   5.858782  5.713194  4.654576  4.792043  7.386009  6.344011  5.740506
#> item 10  6.359140  6.519868  6.226253  7.093199  7.933930  6.713593  8.054612
#> item 11 11.512609  6.376044 11.502417 11.634465  6.749914 11.461080 11.671445
#> item 12 11.538174  6.354960  6.741973  6.418011 11.530548 11.622188 11.603220
#> item 13  6.330339  3.839737  3.822262  3.990550  5.748591  6.402797  6.004828
#> item 14  5.132207  3.633435  3.086874  3.658142  4.818689  5.225482  4.818454
#> item 15  5.380013  4.426031  4.766321  4.387786  5.127774  5.353639  5.074417
#> item 16  6.573114  3.850395  4.208478  4.167015  6.503793  6.595944  6.663296
#> item 17  6.204606  7.424443  7.650946  5.579625  7.542292  8.544501  6.438509
#> item 18  5.488626  9.223433  9.295840  5.186586  9.238719  9.324354  5.997775
#> item 19  7.041727  9.575667  6.637693  9.504269 10.626007  7.675357  9.623505
#> item 20  4.883998  8.532811  8.614262  4.909266  8.450907  8.663136  4.883410
#>              0111      1111
#> item 1   6.859112 12.771521
#> item 2   8.749377  8.630208
#> item 3  11.071170 11.310109
#> item 4  12.163319 12.172053
#> item 5   8.016842  8.136382
#> item 6  13.453244 13.612043
#> item 7  10.663114 10.715242
#> item 8   6.491115  8.285429
#> item 9   6.074754  7.418073
#> item 10  7.313084  8.708345
#> item 11 11.670328 11.660831
#> item 12  7.407881 11.786779
#> item 13  4.676216  6.058610
#> item 14  3.718746  5.131611
#> item 15  4.823958  5.308987
#> item 16  4.081800  6.855066
#> item 17  8.398387  8.398041
#> item 18  9.264553  9.629526
#> item 19  9.630452 10.106430
#> item 20  8.343755  8.322236
# }